# docs/swagger.yaml
basePath: https://api.web95.tech/api/v1
definitions:
  helpers.ErrorData:
    properties:
      code:
        type: string
      fields:
        additionalProperties:
          type: string
        type: object
      message:
        type: string
    type: object
  models.AuthResponse:
    properties:
      token:
        type: string
    type: object
...
  models.HotelRequest:
    properties:
      city:
        type: string
      distance:
        type: number
      distance_text:
        type: string
      guests:
        type: string
      meals:
        type: string
      name:
        type: string
      stars:
        type: integer
      transfer:
        type: string
      urls:
        description: "\U0001F448 массив ссылок"
        items:
          type: string
        type: array
    type: object
  models.HotelResponse:
    properties:
      city:
        type: string
      created_at:
        type: string
      distance:
        type: number
      distance_text:
        type: string
      guests:
        type: string
      id:
        type: integer
      meals:
        type: string
      name:
        type: string
      nights:
        type: integer
      stars:
        type: integer
      transfer:
        type: string
      updated_at:
        type: string
      urls:
        description: "\U0001F448 массив ссылок"
        items:
          type: string
        type: array
    type: object
...
  models.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
...
  models.RegisterRequest:
    properties:
      email:
        type: string
      full_name:
        type: string
      password:
        minLength: 6
        type: string
    required:
    - email
    - password
    type: object
...
  models.TripHotel:
    properties:
      hotel_id:
        type: integer
      nights:
        type: integer
      trip_id:
        type: integer
    type: object
...
  models.User:
    properties:
      avatar:
        type: string
      created_at:
        type: string
      email:
        type: string
      full_name:
        type: string
      id:
        type: integer
      role_id:
        type: integer
      updated_at:
        type: string
    type: object
...
info:
  contact: {}
  description: API for Travel project
  title: Travel API
  version: "1.0"
paths:
...
  /admin/hotels:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.HotelResponse'
            type: array
        "500":
          description: Не удалось получить список отелей
          schema:
            $ref: '#/definitions/helpers.ErrorData'
      summary: List hotels
      tags:
      - Admin — Hotels
    post:
      consumes:
      - application/json
      parameters:
      - description: Hotel
        in: body
        name: hotel
        required: true
        schema:
          $ref: '#/definitions/models.HotelRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.HotelResponse'
        "400":
          description: Некорректные данные
          schema:
            $ref: '#/definitions/helpers.ErrorData'
        "500":
          description: Не удалось создать отель
          schema:
            $ref: '#/definitions/helpers.ErrorData'
      summary: Create hotel
      tags:
      - Admin — Hotels
  /admin/hotels/{id}:
    delete:
      parameters:
      - description: Hotel ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Некорректный ID
          schema:
            $ref: '#/definitions/helpers.ErrorData'
        "500":
          description: Не удалось удалить отель
          schema:
            $ref: '#/definitions/helpers.ErrorData'
      summary: Delete hotel
      tags:
      - Admin — Hotels
    get:
      parameters:
      - description: Hotel ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.HotelResponse'
        "404":
          description: Отель не найден
          schema:
            $ref: '#/definitions/helpers.ErrorData'
      summary: Get hotel by ID
      tags:
      - Admin — Hotels
    put:
      consumes:
      - application/json
      parameters:
      - description: Hotel ID
        in: path
        name: id
        required: true
        type: integer
      - description: Hotel
        in: body
        name: hotel
        required: true
        schema:
          $ref: '#/definitions/models.HotelRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.HotelResponse'
        "400":
          description: Некорректные данные
          schema:
            $ref: '#/definitions/helpers.ErrorData'
        "500":
          description: Не удалось обновить отель
          schema:
            $ref: '#/definitions/helpers.ErrorData'
      summary: Update hotel
      tags:
      - Admin — Hotels
...
  /admin/trips/{id}/hotels:
    post:
      consumes:
      - application/json
      parameters:
      - description: Trip ID
        in: path
        name: id
        required: true
        type: integer
      - description: Hotel ID and Nights
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.TripHotel'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Некорректные данные
          schema:
            $ref: '#/definitions/helpers.ErrorData'
        "500":
          description: Не удалось привязать отель
          schema:
            $ref: '#/definitions/helpers.ErrorData'
      summary: Attach hotel to trip
      tags:
      - Admin — Trips
...
  /auth/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: login payload
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/models.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.AuthResponse'
        "400":
          description: Некорректный запрос
          schema:
            $ref: '#/definitions/helpers.ErrorData'
        "401":
          description: Неверный email или пароль
          schema:
            $ref: '#/definitions/helpers.ErrorData'
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/helpers.ErrorData'
      summary: Login
      tags:
      - System — Auth
  /auth/register:
    post:
      consumes:
      - application/json
      parameters:
      - description: register payload
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/models.RegisterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Некорректные данные
          schema:
            $ref: '#/definitions/helpers.ErrorData'
        "409":
          description: Email уже зарегистрирован
          schema:
            $ref: '#/definitions/helpers.ErrorData'
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/helpers.ErrorData'
      summary: Register
      tags:
      - System — Auth
...
securityDefinitions:
  Bearer:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"